TrainFutures: Fix import paths for moved modules
